+++ terminado/formats/__init__.py
@@ -1,0 +1,1 @@
+"""Message formats for reading and writing to the WebSocket."""